// src/data/bookList.generated.ts
export const TELUGU_BOOKS: Record<string, () => { book: string; chapters: [number, string][][] }> = {
  "gen": () => require("./bible-text/telugu/gen.json"),
  "exo": () => require("./bible-text/telugu/exo.json"),
  "lev": () => require("./bible-text/telugu/lev.json"),
  "num": () => require("./bible-text/telugu/num.json"),
  "deu": () => require("./bible-text/telugu/deu.json"),
  "jos": () => require("./bible-text/telugu/jos.json"),
  "jdg": () => require("./bible-text/telugu/jdg.json"),
  "rut": () => require("./bible-text/telugu/rut.json"),
  "1sa": () => require("./bible-text/telugu/1sa.json"),
  "2sa": () => require("./bible-text/telugu/2sa.json"),
  "1ki": () => require("./bible-text/telugu/1ki.json"),
  "2ki": () => require("./bible-text/telugu/2ki.json"),
  "1ch": () => require("./bible-text/telugu/1ch.json"),
  "2ch": () => require("./bible-text/telugu/2ch.json"),
  "ezr": () => require("./bible-text/telugu/ezr.json"),
  "neh": () => require("./bible-text/telugu/neh.json"),
  "est": () => require("./bible-text/telugu/est.json"),
  "job": () => require("./bible-text/telugu/job.json"),
  "psa": () => require("./bible-text/telugu/psa.json"),
  "pro": () => require("./bible-text/telugu/pro.json"),
  "ecc": () => require("./bible-text/telugu/ecc.json"),
  "sng": () => require("./bible-text/telugu/sng.json"),
  "isa": () => require("./bible-text/telugu/isa.json"),
  "jer": () => require("./bible-text/telugu/jer.json"),
  "lam": () => require("./bible-text/telugu/lam.json"),
  "ezk": () => require("./bible-text/telugu/ezk.json"),
  "dan": () => require("./bible-text/telugu/dan.json"),
  "hos": () => require("./bible-text/telugu/hos.json"),
  "jol": () => require("./bible-text/telugu/jol.json"),
  "amo": () => require("./bible-text/telugu/amo.json"),
  "oba": () => require("./bible-text/telugu/oba.json"),
  "jon": () => require("./bible-text/telugu/jon.json"),
  "mic": () => require("./bible-text/telugu/mic.json"),
  "nam": () => require("./bible-text/telugu/nam.json"),
  "hab": () => require("./bible-text/telugu/hab.json"),
  "zep": () => require("./bible-text/telugu/zep.json"),
  "hag": () => require("./bible-text/telugu/hag.json"),
  "zec": () => require("./bible-text/telugu/zec.json"),
  "mal": () => require("./bible-text/telugu/mal.json"),
  "mat": () => require("./bible-text/telugu/mat.json"),
  "mrk": () => require("./bible-text/telugu/mrk.json"),
  "luk": () => require("./bible-text/telugu/luk.json"),
  "jhn": () => require("./bible-text/telugu/jhn.json"),
  "act": () => require("./bible-text/telugu/act.json"),
  "rom": () => require("./bible-text/telugu/rom.json"),
  "1co": () => require("./bible-text/telugu/1co.json"),
  "2co": () => require("./bible-text/telugu/2co.json"),
  "gal": () => require("./bible-text/telugu/gal.json"),
  "eph": () => require("./bible-text/telugu/eph.json"),
  "php": () => require("./bible-text/telugu/php.json"),
  "col": () => require("./bible-text/telugu/col.json"),
  "1th": () => require("./bible-text/telugu/1th.json"),
  "2th": () => require("./bible-text/telugu/2th.json"),
  "1ti": () => require("./bible-text/telugu/1ti.json"),
  "2ti": () => require("./bible-text/telugu/2ti.json"),
  "tit": () => require("./bible-text/telugu/tit.json"),
  "phm": () => require("./bible-text/telugu/phm.json"),
  "heb": () => require("./bible-text/telugu/heb.json"),
  "jas": () => require("./bible-text/telugu/jas.json"),
  "1pe": () => require("./bible-text/telugu/1pe.json"),
  "2pe": () => require("./bible-text/telugu/2pe.json"),
  "1jn": () => require("./bible-text/telugu/1jn.json"),
  "2jn": () => require("./bible-text/telugu/2jn.json"),
  "3jn": () => require("./bible-text/telugu/3jn.json"),
  "jud": () => require("./bible-text/telugu/jud.json"),
  "rev": () => require("./bible-text/telugu/rev.json"),
};

export const ENGLISH_BOOKS: Record<string, () => { book: string; chapters: [number, string][][] }> = {
  "gen": () => require("./bible-text/english/gen.json"),
  "exo": () => require("./bible-text/english/exo.json"),
  "lev": () => require("./bible-text/english/lev.json"),
  "num": () => require("./bible-text/english/num.json"),
  "deu": () => require("./bible-text/english/deu.json"),
  "jos": () => require("./bible-text/english/jos.json"),
  "jdg": () => require("./bible-text/english/jdg.json"),
  "rut": () => require("./bible-text/english/rut.json"),
  "1sa": () => require("./bible-text/english/1sa.json"),
  "2sa": () => require("./bible-text/english/2sa.json"),
  "1ki": () => require("./bible-text/english/1ki.json"),
  "2ki": () => require("./bible-text/english/2ki.json"),
  "1ch": () => require("./bible-text/english/1ch.json"),
  "2ch": () => require("./bible-text/english/2ch.json"),
  "ezr": () => require("./bible-text/english/ezr.json"),
  "neh": () => require("./bible-text/english/neh.json"),
  "est": () => require("./bible-text/english/est.json"),
  "job": () => require("./bible-text/english/job.json"),
  "psa": () => require("./bible-text/english/psa.json"),
  "pro": () => require("./bible-text/english/pro.json"),
  "ecc": () => require("./bible-text/english/ecc.json"),
  "sng": () => require("./bible-text/english/sng.json"),
  "isa": () => require("./bible-text/english/isa.json"),
  "jer": () => require("./bible-text/english/jer.json"),
  "lam": () => require("./bible-text/english/lam.json"),
  "ezk": () => require("./bible-text/english/ezk.json"),
  "dan": () => require("./bible-text/english/dan.json"),
  "hos": () => require("./bible-text/english/hos.json"),
  "jol": () => require("./bible-text/english/jol.json"),
  "amo": () => require("./bible-text/english/amo.json"),
  "oba": () => require("./bible-text/english/oba.json"),
  "jon": () => require("./bible-text/english/jon.json"),
  "mic": () => require("./bible-text/english/mic.json"),
  "nam": () => require("./bible-text/english/nam.json"),
  "hab": () => require("./bible-text/english/hab.json"),
  "zep": () => require("./bible-text/english/zep.json"),
  "hag": () => require("./bible-text/english/hag.json"),
  "zec": () => require("./bible-text/english/zec.json"),
  "mal": () => require("./bible-text/english/mal.json"),
  "mat": () => require("./bible-text/english/mat.json"),
  "mrk": () => require("./bible-text/english/mrk.json"),
  "luk": () => require("./bible-text/english/luk.json"),
  "jhn": () => require("./bible-text/english/jhn.json"),
  "act": () => require("./bible-text/english/act.json"),
  "rom": () => require("./bible-text/english/rom.json"),
  "1co": () => require("./bible-text/english/1co.json"),
  "2co": () => require("./bible-text/english/2co.json"),
  "gal": () => require("./bible-text/english/gal.json"),
  "eph": () => require("./bible-text/english/eph.json"),
  "php": () => require("./bible-text/english/php.json"),
  "col": () => require("./bible-text/english/col.json"),
  "1th": () => require("./bible-text/english/1th.json"),
  "2th": () => require("./bible-text/english/2th.json"),
  "1ti": () => require("./bible-text/english/1ti.json"),
  "2ti": () => require("./bible-text/english/2ti.json"),
  "tit": () => require("./bible-text/english/tit.json"),
  "phm": () => require("./bible-text/english/phm.json"),
  "heb": () => require("./bible-text/english/heb.json"),
  "jas": () => require("./bible-text/english/jas.json"),
  "1pe": () => require("./bible-text/english/1pe.json"),
  "2pe": () => require("./bible-text/english/2pe.json"),
  "1jn": () => require("./bible-text/english/1jn.json"),
  "2jn": () => require("./bible-text/english/2jn.json"),
  "3jn": () => require("./bible-text/english/3jn.json"),
  "jud": () => require("./bible-text/english/jud.json"),
  "rev": () => require("./bible-text/english/rev.json"),
};